:building_construction: -Ajout des tables Role et Utilisateur -Amélioration des routes (Rendering, version de l'appli, et session) -Ajout de check des permissions avant d'effectuer les actions -Ajout de l'ellipse dans utils-dt.js -Possibilité de se connecter en cliquant en haut à gauche

+++ routes/action.js
@@ -1,13 +1,16 @@
 var express = require('express');
 var router = express.Router();
 const sequelize = require('sequelize');
-const models = require('./../standalone')();
+const {models : initModels} = require('./../standalone');
 
+models = initModels();
 
 module.exports = () => {
 
     router.use(express.json());
     router.use(express.urlencoded({ extended: true}));
+
+    var session;
 
     var response = {
         status: 'nok',
@@ -23,10 +26,72 @@ module.exports = () => {
         response.message = message;
     }
 
+
+    /**
+     * Vérifie si un utilisateur a la permission d'effectuer une action
+     * @param id_utilisateur
+     * @param droit
+     * @returns {*}
+     */
+    async function checkDroit(id_utilisateur, droit){
+        let r;
+        switch(droit){
+            case 'create':
+                r = await models.Utilisateur.canCreate(id_utilisateur);
+                break
+            case 'update':
+                r = await models.Utilisateur.canUpdate(id_utilisateur);
+                break;
+            case 'delete':
+                r = await models.Utilisateur.canDelete(id_utilisateur);
+                break;
+            default:
+                r = false;
+        }
+        return r;
+    }
+
     router.use(function timeLog (req, res, next) {
         //console.dir(req.url);
         next();
     });
+
+
+    router.post('/isConnected', (req, res, next) => {
+        res.send(!!session);
+    });
+
+    router.post('/login', function(req, res){
+        if(!req.body.login || !req.body.password){
+            response.title = 'Connexion impossible';
+            response.message = 'Veuillez vérifier vos informations';
+            res.send(response);
+
+        } else {
+            //Ici, on vérifie le login/password
+            models.Utilisateur.checkCredentials(req.body.login, req.body.password).then((data) => {
+                if(data){
+                    req.session.user = data.id_utilisateur;
+                    req.session.nom = data.nom;
+                    req.session.picture = data.picture;
+                    req.session.permissions = data.permissions;
+                    session = req.session;
+                    console.log('[User logged]: ' + req.session.nom);
+                    setResponseOk('', '');
+                    res.send(response);
+                }
+            });
+        }
+    })
+
+
+    router.get('/logout', function(req, res){
+        req.session.destroy(function(){
+            console.log("user logged out.")
+        });
+        res.redirect('/login');
+    });
+
 
 
     router.post('/loadTalents', function(req, res) {
@@ -46,6 +111,11 @@ module.exports = () => {
     });
 
     router.post('/createTalent', function(req, res) {
+        if(!checkDroit(req.session.user, 'create')){
+            response.message = 'Permission non accordée.';
+            res.send(response);
+            return false;
+        }
         models.Talent.createTalent(req.body).then((data) => {
             if(data){
                 setResponseOk('Création réussie', 'Talent ' + data.nom + ' créé !');
@@ -55,6 +125,12 @@ module.exports = () => {
     });
 
     router.post('/updateTalent', function(req, res) {
+        if(!checkDroit(req.session.user, 'update')){
+            response.message = 'Permission non accordée.';
+            res.send(response);
+            return false;
+        }
+
         let id_talent = req.body.id_talent;
         delete req.body.id_talent;
 
@@ -67,6 +143,12 @@ module.exports = () => {
     });
 
     router.post('/deleteTalent', function(req, res) {
+        if(!checkDroit(req.session.user, 'delete')){
+            response.message = 'Permission non accordée.';
+            res.send(response);
+            return false;
+        }
+
         models.Talent.deleteTalent(req.body.id_talent).then((data) => {
             if(data){
                 setResponseOk('Suppression réussie', 'Le talent a été supprimé !');
